requirements.txt
streamlit
pdfplumber
python-docx
openpyxl
pandas
